move swerve constants from project-wide file to local
