+++ src/main/java/frc/robot/subsystems/swerve/GyroIOPigeon2.java
@@ -6,7 +6,6 @@ import com.ctre.phoenix6.configs.Pigeon2Configuration;
 import com.ctre.phoenix6.hardware.Pigeon2;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.util.Units;
-import frc.robot.Constants.Swerve;
 
 public class GyroIOPigeon2 implements GyroIO {
   private final Pigeon2 pigeon;
@@ -14,7 +13,7 @@ public class GyroIOPigeon2 implements GyroIO {
   private final StatusSignal<Double> yawVelocity;
 
   public GyroIOPigeon2() {
-    pigeon = new Pigeon2(Swerve.GYRO_ID);
+    pigeon = new Pigeon2(DriveConstants.GYRO_ID);
 
     pigeon.getConfigurator().apply(new Pigeon2Configuration());
     pigeon.setYaw(0);
